perf: add more validattion

+++ controllers/search-controller.js
@@ -1,13 +1,21 @@
 const searchServices = require("../services/Search-service");
+const createError = require("../utils/createError");
 
 const searchController = {};
 
 searchController.getSearch = async(req,res,next) => {
     try {
         const userId = req.user.id;
-        console.log(req.user)       
+        if(!userId) {
+            return createError(400,'user ID is require')
+        }     
+
         const getUserSearch = await searchServices.getSearch(userId);
-        res.status(200).json({getUserSearch})
+        if(!getUserSearch ||getUserSearch.length === 0) {
+            return createError(404,'no search history found')
+        }
+
+        res.status(200).json(getUserSearch)
     }catch(err) {
         console.log('error form getSearch',err)
         next(err)
@@ -16,10 +24,17 @@ searchController.getSearch = async(req,res,next) => {
 searchController.createSearch = async(req,res,next) => {
     try {
         const userId = req.user.id;
+        if(!userId) {
+            return createError(400,'user ID is require')
+        }
 
         const {searchTerm} = req.body;
+        if(!searchTerm) {
+            return createError(400, 'searchTerm is require')
+        }
+
         const createSearch = await searchServices.createSearch(userId,searchTerm)
-        res.status(201).json({createSearch});
+        res.status(201).json(createSearch);
     } catch (err) {
         console.log('error form createSearch',err)
         next(err)
@@ -28,11 +43,15 @@ searchController.createSearch = async(req,res,next) => {
 searchController.deleteSearch = async(req,res,next) => {
     try {
         const userId = req.user.id;
+        if(!userId) {
+            return createError(400,'user ID is require')
+        }
+
         const { historyId } = req.params;
-        console.log(req.params)
         if (!historyId) {
             return res.status(400).json({ message: 'Search ID is required' });
         }
+        
         const search = await searchServices.deleteSearch(userId,historyId);
         res.status(204).json({message: 'Search history deleted successfully'});
     }catch (err) {
